Fix: Refactor moviefile data, event handling
