+++ frontend/src/App/State/MovieFilesAppState.ts
@@ -1,10 +1,12 @@
 import AppSectionState, {
   AppSectionDeleteState,
+  TableAppSectionState,
 } from 'App/State/AppSectionState';
 import { MovieFile } from 'MovieFile/MovieFile';
 
 interface MovieFilesAppState
   extends AppSectionState<MovieFile>,
-    AppSectionDeleteState {}
+    AppSectionDeleteState,
+    TableAppSectionState {}
 
 export default MovieFilesAppState;
