+++ frontend/src/Components/SignalRConnector.js
@@ -333,17 +333,8 @@ class SignalRConnector extends Component {
   };
 
   handleMoviefile = (body) => {
-    const section = 'movieFiles';
-
-    if (body.action === 'updated') {
-      this.props.dispatchUpdateItem({ section, ...body.resource });
-
-      // Repopulate the page to handle recently imported file
-      repopulatePage('movieFileUpdated');
-    } else if (body.action === 'deleted') {
-      this.props.dispatchRemoveItem({ section, id: body.resource.id });
-
-      repopulatePage('movieFileDeleted');
+    if (body.action === 'updated' || body.action === 'deleted') {
+      queryClient.invalidateQueries({ queryKey: ['/moviefile'] });
     }
   };
